Add default impl for validate()
